+++ src/main/java/io/appform/dropwizard/actors/base/UnmanagedConsumer.java
@@ -8,19 +8,18 @@ import io.appform.dropwizard.actors.actor.ActorConfig;
 import io.appform.dropwizard.actors.actor.ConsumerConfig;
 import io.appform.dropwizard.actors.actor.MessageHandlingFunction;
 import io.appform.dropwizard.actors.base.utils.NamingUtils;
+import io.appform.dropwizard.actors.common.RabbitmqActorException;
 import io.appform.dropwizard.actors.connectivity.RMQConnection;
 import io.appform.dropwizard.actors.exceptionhandler.ExceptionHandlingFactory;
 import io.appform.dropwizard.actors.exceptionhandler.handlers.ExceptionHandler;
 import io.appform.dropwizard.actors.observers.RMQObserver;
 import io.appform.dropwizard.actors.retry.RetryStrategy;
 import io.appform.dropwizard.actors.retry.RetryStrategyFactory;
-
-import java.util.Optional;
-import lombok.extern.slf4j.Slf4j;
-
 import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
-
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 
 @Slf4j
@@ -39,8 +38,8 @@ public class UnmanagedConsumer<Message> {
     private final RetryStrategy retryStrategy;
     private final ExceptionHandler exceptionHandler;
     private final RMQObserver observer;
-
     private final List<Handler<Message>> handlers = Lists.newArrayList();
+    private int consumerIndex = 1;
 
     public UnmanagedConsumer(final String name,
                              final ActorConfig config,
@@ -69,21 +68,13 @@ public class UnmanagedConsumer<Message> {
 
     public void start() throws Exception {
         for (int i = 1; i <= config.getConcurrency(); i++) {
-            Channel consumeChannel = connection.newChannel();
-            final Handler<Message> handler =
-                    new Handler<>(consumeChannel, mapper, clazz, prefetchCount, errorCheckFunction, retryStrategy,
-                                  exceptionHandler, handlerFunction, expiredMessageHandlingFunction, observer, queueName);
             String queueNameForConsumption;
             if (config.isSharded()) {
                 queueNameForConsumption = NamingUtils.getShardedQueueName(queueName, i % config.getShardCount());
             } else {
                 queueNameForConsumption = queueName;
             }
-
-            final String tag = consumeChannel.basicConsume(queueNameForConsumption, false, getConsumerTag(i), handler);
-            handler.setTag(tag);
-            handlers.add(handler);
-            log.info("Started consumer {} of type {} with tag {}", i, name, tag);
+            createChannel(queueNameForConsumption);
         }
     }
 
@@ -95,7 +86,9 @@ public class UnmanagedConsumer<Message> {
                     channel.basicCancel(handler.getTag());
                     //Wait till the handler completes consuming and ack'ing the current message.
                     log.info("Waiting for handler to complete processing the current message..");
-                    while(handler.isRunning());
+                    while (handler.isRunning()) {
+                        // wait for the handler to complete processing the current message
+                    }
                     channel.close();
                     log.info("Consumer channel closed for [{}] with prefix [{}]", name, config.getPrefix());
                 } else {
@@ -119,4 +112,31 @@ public class UnmanagedConsumer<Message> {
     public String queueName() {
         return queueName;
     }
+
+    private void channelClosedHandler(String queueName) {
+        try {
+            log.info("Re-creating channel for queue {} because a channel got closed", queueName);
+            createChannel(queueName);
+        } catch (Exception e) {
+            throw RabbitmqActorException.propagate(e);
+        }
+    }
+
+    private void createChannel(String queueName) throws Exception {
+        Channel consumeChannel = connection.newChannel();
+        consumeChannel.addShutdownListener(shutdownSignalException -> {
+            // This needs to be run in a separate thread to avoid deadlock
+            CompletableFuture.runAsync(() -> {
+                log.error("Channel shutdown signal exception received for queue {}", queueName, shutdownSignalException);
+                channelClosedHandler(queueName);
+            });
+        });
+        final Handler<Message> handler = new Handler<>(consumeChannel, mapper, clazz, prefetchCount, errorCheckFunction,
+                retryStrategy, exceptionHandler, handlerFunction, expiredMessageHandlingFunction, observer, queueName);
+        final String tag = consumeChannel.basicConsume(queueName, false, getConsumerTag(consumerIndex++), handler);
+        handler.setTag(tag);
+        handlers.add(handler);
+        log.info("Started consumer for queue {} with tag {}", name, tag);
+    }
+
 }